requirements.txt
annotated-types==0.7.0
anyio==4.6.0
certifi==2024.8.30
charset-normalizer==3.4.0
distro==1.9.0
h11==0.14.0
httpcore==1.0.6
httpx==0.27.2
idna==3.10
jaraco.classes==3.4.0
jaraco.context==6.0.1
jaraco.functools==4.1.0
jiter==0.6.1
keyring==25.4.1
lxml==5.3.0
more-itertools==10.5.0
openai==1.51.2
pydantic==2.9.2
pydantic_core==2.23.4
python-docx==1.1.2
PyYAML==6.0.2
regex==2024.9.11
requests==2.32.3
sniffio==1.3.1
tiktoken==0.8.0
tqdm==4.66.5
typing_extensions==4.12.2
urllib3==2.2.3
